+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/Platform.java
@@ -66,6 +66,8 @@ public interface Platform {
 
     Optional<FluidResource> convertToFluid(ItemStack stack);
 
+    Optional<FluidResource> convertToFluid(Object value);
+
     EnergyStorage createEnergyStorage(ControllerType controllerType, Runnable listener);
 
     void setEnergy(EnergyStorage energyStorage, long stored);