+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/packet/ClientToServerCommunications.java
@@ -5,6 +5,7 @@ import com.refinedmods.refinedstorage2.api.grid.service.GridInsertMode;
 import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
 import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
+import com.refinedmods.refinedstorage2.platform.api.resource.filter.FilteredResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.filter.ResourceType;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.PropertyType;
 
@@ -22,6 +23,8 @@ public interface ClientToServerCommunications {
     <T> void sendPropertyChange(PropertyType<T> type, T value);
 
     void sendResourceTypeChange(ResourceType type);
+
+    void sendResourceFilterSlotChange(int slotIndex, FilteredResource filteredResource);
 
     void sendStorageInfoRequest(UUID storageId);