+++ refinedstorage2-platform-common/src/test/java/com/refinedmods/refinedstorage2/platform/common/internal/storage/PlatformStorageRepositoryImplTest.java
@@ -134,6 +134,7 @@ class PlatformStorageRepositoryImplTest {
     }
 
     @Test
+    @SuppressWarnings("unchecked")
     void shouldSerializeAndDeserialize() {
         // Arrange
         final InMemoryTrackedStorageRepository<ItemResource> repository = new InMemoryTrackedStorageRepository<>();