+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/packet/c2s/ClientToServerCommunicationsImpl.java
@@ -6,6 +6,7 @@ import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
 import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
+import com.refinedmods.refinedstorage2.platform.api.resource.filter.FilteredResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.filter.ResourceType;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.PropertyType;
 import com.refinedmods.refinedstorage2.platform.common.packet.ClientToServerCommunications;
@@ -52,6 +53,11 @@ public class ClientToServerCommunicationsImpl implements ClientToServerCommunica
         PlatformApi.INSTANCE.getResourceTypeRegistry()
             .getId(type)
             .ifPresent(id -> networkManager.send(new ResourceTypeChangePacket(id)));
+    }
+
+    @Override
+    public void sendResourceFilterSlotChange(final int slotIndex, final FilteredResource filteredResource) {
+        networkManager.send(new ResourceFilterSlotChangePacket(slotIndex, filteredResource));
     }
 
     @Override